backend/requirements.txt
# requirements.txt for TravelTrails backend

# FastAPI and related
fastapi>=0.100.0 # 建议升级，或者降级 starlette
uvicorn[standard]
pydantic>=1.10.0,<2.0.0 # 兼容旧版.dict()
python-dotenv

# For AI Recommendation Service
httpx
geopy

# For User Management (password hashing - strongly recommended)
# passlib[bcrypt]

# Database (SQLAlchemy - if/when migrating from users.json)
SQLAlchemy # 示例中DAO用，如果不用数据库则可选
psycopg2-binary # 如果用PostgreSQL
# cryptography # 如果要自己处理密码哈希，但 werkzeug.security 更常用

# JWT for authentication (if implementing token-based auth)
# python-jose[cryptography]

# Example versions (update as needed):
# fastapi==0.104.1
# uvicorn[standard]==0.23.2
# pydantic==2.4.2
# python-dotenv==1.0.0
# httpx==0.25.0
# geopy==2.4.0
# passlib[bcrypt]==1.7.4
# sqlalchemy==2.0.22
# databases[sqlite]==0.8.0
# python-jose[cryptography]==3.3.0 